perf: Improve animated tile images
The previous implementation "leaked" quite a bit of memory. It wasn't a
proper memory leak in that the memory would still eventually be freed, but
it did end up allocating more memory than would be desirable.

The implementation has been rewritten in a bit more of an efficient way.
Now, instead of new bitmaps being created for each frame each time said
frame needs to be rendered, the entire strip of bitmaps will be loaded into
memory once. The implementation details are quite simple: the sprite sheet
has its position translated within a canvas, and then that canvas has its
contents clipped to effectively render only one animation frame.

To facilitate this change, the control was moved from a manual Control to
a UserControl with associated XAML. I doubt that this was *necessary*, per
se, but it does result in cleaner code than if this choice were avoided.

The class has been renamed from AnimatedTileSequence in C# contexts and
TileImage in XAML contexts to share the name AnimatedTileImage in both
contexts.
