+++ Aerochat/Controls/AnimatedTileImage.xaml
@@ -1,0 +1,16 @@
+﻿<!-- Play an image from an animated sprite sheet. -->
+<UserControl x:Class="Aerochat.Controls.AnimatedTileImage"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Aerochat.Controls"
+             DataContext="{Binding RelativeSource={RelativeSource self}}"
+             mc:Ignorable="d"
+             Width="{Binding FrameWidth}"
+             Height="{Binding FrameHeight}"
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Canvas ClipToBounds="True">
+        <Image x:Name="_imageElement" Source="{Binding Image}" Width="{Binding Image.Width}" Height="{Binding Image.Height}" SnapsToDevicePixels="True"/>
+    </Canvas>
+</UserControl>
