+++ Aerochat/Windows/Home.xaml
@@ -224,7 +224,7 @@
                                 </Style.Triggers>
                             </Style>
                         </Grid.Resources>
-                        <controls:TileImage PreviewMouseLeftButtonUp="SceneTileImage_PreviewMouseLeftButtonUp" Cursor="Hand" x:Name="SceneTileImage" HorizontalAlignment="Right" VerticalAlignment="Top" FrameWidth="31" FrameHeight="31" FrameDuration="75" Width="31" Height="31" Loop="False" />
+                        <controls:AnimatedTileImage PreviewMouseLeftButtonUp="SceneTileImage_PreviewMouseLeftButtonUp" Cursor="Hand" x:Name="SceneTileImage" HorizontalAlignment="Right" VerticalAlignment="Top" FrameWidth="31" FrameHeight="31" FrameDuration="75" Width="31" Height="31" Loop="False" />
                         <Border Style="{StaticResource Shadow}" VerticalAlignment="Center" HorizontalAlignment="Stretch" Margin="80,-10,0,0">
                             <Border Style="{StaticResource Shadow}">
                                 <Grid>
